refactored routes and models, added ECR logic to CI
