+++ backend/v1/app/database/__init__.py
@@ -1,0 +1,1 @@
+"""CurlUp Backend System Database Module."""
