+++ backend/v1/app/models/admin.py
@@ -1,0 +1,1 @@
+"""CurlUp Backend System Admin Models."""
