+++ backend/v1/app/routes/admin.py
@@ -1,0 +1,1 @@
+"""CurlUp Backend Administrator Routes."""
