--- backend/v1/app/routes/system/home.py
@@ -1,11 +1,0 @@
-"""CurlUp Home Endpoint"""
-
-from fastapi import APIRouter
-
-router = APIRouter()
-
-
-@router.get("/")
-async def home():
-    """Home Endpoint."""
-    return {"message": "Welcome to the CurlUp API"}
