+++ backend/v1/app/schema/client.py
@@ -1,0 +1,1 @@
+"""CurlUp Backend Client Tables Schemas."""
