--- backend/v1/app/schema/users.py
@@ -1,1 +1,0 @@
-"""CurlUp Backend Users Table Schema."""
